Squashed commit of the following:

commit 9efcb7ae90061cf01696eafdf78fa36b93ba0dde
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 13:29:03 2025 +1000

    version

commit a0d4ff74673c0c668badc611d49fb7bc544ab4e2
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 13:23:27 2025 +1000

    ~ formatting

commit 5c036ecdf135b8f392205761d4b2dfd4216c316c
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 13:12:59 2025 +1000

    + added and applied `unixem_util_fs_directory_exists()`

commit 2a025bbf8d1fe54c55e924cd6c44963237deef0b
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 12:56:13 2025 +1000

    `glob()` : ~ refactored in terms of `unixem_util_fs_get_home_directory()`

commit a41c456f1aa65fd6c18e2d6b8329a50796049cf8
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 12:54:46 2025 +1000

    ~ fix

commit 3e690500634ce1ebddfe1f934dc4d81e9b2455b0
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 12:11:53 2025 +1000

    `glob()` : ~ preparatory refactoring

commit e6ee442cf4ac93944d29d30a2a703f8e436f8c9d
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 12:04:46 2025 +1000

    + added `unixem_util_fs_get_home_directory()`

commit bc575352d77f581c77fb71a461b7899d4304d1b4
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sun May 18 12:01:37 2025 +1000

    + added **UNIXem.Util.Str** API

commit 25e1376ff0ce657e09d281c9f7a385e27a7da7b1
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Sat May 17 16:16:52 2025 +1000

    ~ preparatory refactoring

commit 0a95c752583ebb4880424b551d8bb0014bf0ee57
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Tue May 13 17:10:07 2025 +1000

    `glob()` : ~ simplifying refactoring

commit cb3b89645fbdf19ec9da92402e4c7822515e46e4
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Tue May 13 17:05:37 2025 +1000

    `glob()` : ~ simplifying refactoring

commit cefaacb9e3df59a16c0cc210122090c83070ba72
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Tue May 13 16:59:32 2025 +1000

    + added **test.unit.util.fs**

commit 1c10a4b046f3f0b8864fd9a75e7aac176c3bf694
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Tue May 13 16:29:20 2025 +1000

    **test.component.unistd.mkdtemp** / **test.component.unistd.mkdtemp** : ~ minor mods

commit 6a56dabc645949d0e958c32b8a7f79970dd7b27f
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Mon May 12 08:39:09 2025 +1000

    TODO.md : + more items

commit 6d5edb750ccd70621626aced966ca94ed1e9d2fa
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Mon May 12 08:38:00 2025 +1000

    ~ renamed TODO.txt => TODO.md

commit 045cc3251a3f5561580b1eddcb3cc6202b6ee749
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Mon May 12 08:12:39 2025 +1000

    **UNIXem.Util.FS** : + added `unixem_util_fs_char_is_path_sep()`

commit 279349e5fa3f9812ea37ea5860f42af821cb51ed
Author: Matt Wilson <matthew@synesis.com.au>
Date:   Mon May 12 08:05:41 2025 +1000

    + added **UNIXem.Util.FS** implementation file to src/CMakeLists.txt

+++ test/component/test.component.unistd.mkdtemp/entry.c
@@ -1,14 +1,17 @@
 
 /*
- * Created:
- * Updated: 29th November 2024
+ * Created: ...
+ * Updated: 18th May 2025
  */
 
 #include <unistd.h>
 
-#include <xtests/xtests.h>
+#include <xtests/terse-api.h>
+
+#include <unixem/util/fs.h>
 
 #include <stdlib.h>
+
 
 int main(int argc, char** argv)
 {
@@ -29,7 +32,7 @@ int main(int argc, char** argv)
             {
                 int const e = errno;
 
-                fprintf(stderr, "mkdtemp() failed: %s (%d)\n", strerror(e), e);
+                fprintf(stderr, "mkdtemp() failed to create '%s': %s (%d)\n", template_path, strerror(e), e);
             }
             else
             {
@@ -50,27 +53,30 @@ int main(int argc, char** argv)
         /* Test-2 */
         if (XTESTS_CASE_BEGIN("Test-2", "testing mkdtemp in specific directory"))
         {
-            char        template_path[] =   "C:\\temp\\abc.XXXX";
-            char const* r               =   mkdtemp(template_path);
-
-            if (NULL == r)
+            if (unixem_util_fs_directory_exists("C:\\temp"))
             {
-                int const e = errno;
+                char        template_path[] =   "C:\\temp\\abc.XXXX";
+                char const* r               =   mkdtemp(template_path);
 
-                fprintf(stderr, "mkdtemp() failed: %s (%d)\n", strerror(e), e);
-            }
-            else
-            {
-                size_t const len = strlen(template_path);
+                if (NULL == r)
+                {
+                    int const e = errno;
 
-                rmdir(template_path);
+                    fprintf(stderr, "mkdtemp() failed to create '%s': %s (%d)\n", template_path, strerror(e), e);
+                }
+                else
+                {
+                    size_t const len = strlen(template_path);
 
-                XTESTS_REQUIRE(XTESTS_TEST_INTEGER_EQUAL(16u, len));
-                XTESTS_TEST_MULTIBYTE_STRING_EQUAL_N("C:\\temp\\abc.", template_path, 12u);
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 4]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 3]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 2]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 1]));
+                    rmdir(template_path);
+
+                    XTESTS_REQUIRE(XTESTS_TEST_INTEGER_EQUAL(16u, len));
+                    XTESTS_TEST_MULTIBYTE_STRING_EQUAL_N("C:\\temp\\abc.", template_path, 12u);
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 4]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 3]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 2]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 1]));
+                }
             }
 
             XTESTS_CASE_END("Test-2");
